Force include Sphinx’s search model when the plugin is enabled.
There is currently a nasty bug in the autoloader when the object cache isn’t fully built. It causes the default search model to get loaded even when Sphinx is enabled. This leads to database killing queries on larger sites.

This fix is a bit of a kludge, but just makes sure that Sphinx’s search model will always load if the plugin is enabled.