Simplify event timezone handling
Use the browser to determine proper event timezones rather than having the user select the timezone when creating the event.

The date/time picker now maintains a hidden field that synchronizes the proper date/time in a hidden field. There is no need to guess a timezone anymore making things easier for the user and more robust.