Rename $PluginInfo variable to avoid scanPluginFile() conflict
The Gdn_PluginManager::scanPluginFile() looks for a line that starts
with `$PluginInfo` to search for a plugin info array. Since we’ve
removed the info arrays a bug has opened up where subsequent lines with
that variable are falsely flagged as the info array.

See also https://github.com/vanilla/vanilla/pull/6091.