Clear revisions cache after publishing a new article
We don't get the data back from the API that would allow us to purge smaller parts of the cache so I'm clearing it entirely when we patch an article (generating a new revision and modifying the active one).

If we have the need we can do some profiling in the future an d do more selective caching here, but with how much data we preload it's probably not an issue.