Change the article dtypes to be in the 500’s
The dtype attribute in sphinx is “namespaced” by its index so articles being type 5 should all have dtypes in the 500’s. This is to leave space available for the other indexes.

This change has the potential to break searching until the templates are updated properly so there is a kludge to still respect dtype 5.