Refactor Sphinx search for easier support of new types
This is a fairly large refactor of the sphinx and advanced search addons in order to support adding new types.

Consider what needs to be done to search a new type:

1. The API has to be aware of the type and what indexes it uses.
2. Advanced search has to know how to join records into the search results.
3. The advanced search UI must be able to display the types that are available for search.

This was currently being done with various properties on the `Search`, `AdvancedSearchPlugin`, and `SphinxSearchModel` classes and it was quite difficult to add new types because the properties were spread across objects and had opaque, undocumented uses. I had to do a lot of debugging to try and support KB search in our regular advanced search so I decided to refactor the workflow to make it easier for future types. Here are the changes:

1. All of the logic has been moved to the `SphinxSearchModel` so that it can be augmented in the container without requiring any other objects to be touched.
2. Addons can now use `SphinxSearchModel::addDocumentType()` to register document fetching behavior.
3. Addons can now use `SphinxSearchModel::addSearchType()` to register search types for the advanced search module and search endpoint.
4. The advanced search module and APIv1 endpoint now accepts a more sensible array of `types` rather than individual boolean parameters. There is a backwards compatibility layer to support the old parameters.

All the other properties have been deprecated. I don’t think they are used outside of sphinx and advanced search currently, but I left them in to prevent site crashing bugs if I missed something.