Fix the /api/v2/discussions|comments/:id reactions endpoint for idea votes
Getting the individual vote reactions for ideas was failing because ideation turns those reactions off and then activates them behind the scenes. There was already an event to fix this for reactions summaries, but it wasn’t applied to the full list of reactions.

This PR has the following fixes:

1. I added a new `ReactionModel::getRecordReactions()` that is a refactor of `ReactionModel::getByRecord()`. The new method takes an entire row, rather than a `recordID/recordType` pair to avoid a redundant database lookup.

2. The new method fires the event that ideation needs.

3. I added an extra assertion in the ideation test to ensure the reactions are coming back for ideas.

4. I added a bunch of docblock fixes to try and wrangle the yellow bars in my IDE. Not much success on the bars, but I think the docblocks are a good thing here.

Closes https://github.com/vanilla/internal/issues/2067.