Change the reaction types filter
I am giving the event a more semantic name now that is known what its purpose is. I’m also removing the reaction model from the sender so that addons don’t start using it to do things that aren’t the purpose of the event.

I’ve searched all our repos for handlers and ideation seems to be the only one 🤞.