Make external link checking for ranks more robust
The ranks plugin had a few issues with external links.

- Our CDN was being treated as an external link which disallowed image uploads.
- External links rendered with `/entry/leaving` were being treated as internal links.

This PR moves the link checking into the `RankModel` for easier testing. It also adds some tests on the external link functionality.

Closes https://github.com/vanilla/support/issues/126.
Closes https://github.com/vanilla/support/issues/786.

Note that a couple of the issues above were already fixed by another PR so this PR is just addressing external link functionality in general.