Make automatic badges be given by the system user
This might be considered more of an enhancement rather than a fix, because the referenced bug was actually fixed previously. However, I think this is a more consistent way for the user badge table to work.

This change generally makes `Gdn_UserBadge.InsertUserID` default to the system user ID except in the following cases:

1. When a user requests a badge then they are going to be the insert user.
2. When a badge is manually awarded then the giver is going to be the insert user.


Closes #1101