Fix interaction problem between reactions and ideation plugin
Ideation stips the reactions from the discussion.
Since the reaction types are statically cached and were passed by reference this means that reaction types were set to an empty array for the whole page.
This caused comments to not have reactions on the first page.

This PR fixes that by doing a copy of of the reactions types before assigning a reference to the event's argument.