Fix Gdn_SQLDriver::noReset()
There seems to be a bug in `Gdn_SQLDriver::noReset()`. This is one of those things that is definitely a fix, but could also lead to a regression. The only place I saw a usage in our entire code base is in `Gdn_Model` There was actually a typo above that was making that code path be hit in tests too which I fixed.

Finally, some tests were throwing exceptions because of an error in `Gdn_SQLDriver::truncate()` where they pass a non-existent option of ‘truncate’ to `Gdn_Database::query()`.